requirements.txt
sagemaker>=2.200.0
boto3>=1.34.0
transformers>=4.36.0
torch>=2.1.0
datasets>=2.16.0
huggingface-hub>=0.20.0
pyyaml>=6.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
evaluate>=0.4.1
peft>=0.7.0
accelerate>=0.25.0
trl>=0.7.10
